+++ src/peigs/src/c/memreq.c
@@ -178,7 +178,7 @@ void memreq_(type, n, mapA, mapB, mapZ, isize, rsize, ptr_size, iscratch )
      for ( indx = 0; indx < *n; indx++ ) {
        if ( (iptr++ ) == NULL ) {
          linfo = -4;
-         fprintf(stderr, " me = %d error in mapB in memreq.c \n",me);
+         fprintf(stderr, " me = %d error in mapB in memreq.c \n",(int)me);
          l_exit_(&linfo,msg);
          return;
        }
@@ -194,7 +194,7 @@ void memreq_(type, n, mapA, mapB, mapZ, isize, rsize, ptr_size, iscratch )
    for ( indx = 0; indx < *n; indx++ ) {
      if ( (iptr++ ) == NULL ) {
        linfo = -5;
-       fprintf(stderr, " me = %d error in mapZ in memreq.c \n",me);
+       fprintf(stderr, " me = %d error in mapZ in memreq.c \n",(int)me);
        l_exit_(&linfo,msg);
        return;
      }
