+++ src/peigs/src/c/pstein5.c
@@ -178,7 +178,7 @@ void pstein5 ( n, dd, ee, dplus, lplus, ld, lld, meigval, eval, iblock, nsplit, 
   
   
 #ifdef DEBUG1
-  fprintf(stderr, "me = %d In pstein \n", me );
+  fprintf(stderr, "me = %d In pstein \n", (int)me );
 #endif
   
   /*
@@ -230,7 +230,7 @@ void pstein5 ( n, dd, ee, dplus, lplus, ld, lld, meigval, eval, iblock, nsplit, 
     if ( info != NULL )
         *info = linfo;
     printf( " %s me = %d argument %d is a pointer to NULL. \n",
-	    msg, me, -linfo );
+	    msg, (int)me, (int)(-linfo) );
     fflush(stdout);
     xstop_( &linfo );
     return;
@@ -298,7 +298,7 @@ void pstein5 ( n, dd, ee, dplus, lplus, ld, lld, meigval, eval, iblock, nsplit, 
   if( *info != 0 ) {
     linfo = *info;
     printf( " %s me = %d argument %d has an illegal value. \n",
-	    msg, me, -linfo);
+	    msg, (int)me, (int)(-linfo));
     xstop_( info );
     return;
   }
@@ -442,7 +442,7 @@ void pstein5 ( n, dd, ee, dplus, lplus, ld, lld, meigval, eval, iblock, nsplit, 
   if ( isize < 0 ){
     *info = -99;
     fprintf(stderr, " Node %d: error in clustrf isize = %d neigval = %d \n", 
-	    me, isize, neigval );
+	    (int)me, (int)isize, (int)neigval );
     xstop_( info );
   }
